Refactor product form: change category input to a select dropdown and update price field type to decimal

+++ database/migrations/2026_05_09_072319_create_products_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->string('slug')->unique();
             $table->string('category'); 
             $table->text('description')->nullable();
-            $table->unsignedInteger('price'); // rupiah
+            $table->decimal('price', 12, 2);
             $table->string('image')->nullable(); // path storage
             $table->boolean('status')->default(true); // tersedia / tidak
             $table->unsignedInteger('sold_count')->default(0); // produk terlaris (opsional tapi berguna)
